refactor admin routes to be secured
