+++ components/InfoGrid.tsx
@@ -1,27 +1,26 @@
-import { Grid } from '@mui/material';
-import DashboardCard from './DashboardCard';
-import ListCard from './ListCard';
+import { Grid } from "@mui/material";
+import DashboardCard from "./DashboardCard";
 
 const InfoGrid = (props: any) => {
-    const listItems = props.listItems;
-    return (
-        <Grid
-            container
-            rowSpacing={4}
-            columnSpacing={{ xs: 2, sm: 2, md: 3 }}
-            alignContent="stretch"
-            alignItems="stretch"
-        >
-            {listItems.map((listItem: any, index: number) => {
-                return (
-                    <DashboardCard
-                        key={index}
-                        link={listItem.link}
-                        title={listItem.title}
-                    />
-                );
-            })}
-        </Grid>
-    );
+	const listItems = props.listItems;
+	return (
+		<Grid
+			container
+			rowSpacing={4}
+			columnSpacing={{ xs: 2, sm: 1, md: 2 }}
+			alignContent="stretch"
+			alignItems="stretch"
+		>
+			{listItems.map((listItem: any, index: number) => {
+				return (
+					<DashboardCard
+						key={index}
+						link={listItem.link}
+						title={listItem.title}
+					/>
+				);
+			})}
+		</Grid>
+	);
 };
 export default InfoGrid;
