+++ pages/admin/categories/add-category.tsx
