+++ pages/admin/equines/add-equine.tsx
