+++ pages/admin/learner-types/[id].tsx
@@ -11,8 +11,8 @@ import { useRouter } from 'next/router';
 import { useState, useEffect } from 'react';
 import { withRouter, NextRouter } from 'next/router';
 import DeleteForeverIcon from '@mui/icons-material/DeleteForever';
-import PageTitle from '../../components/PageTitle';
-import LinkButton from '../../components/LinkButton';
+import PageTitle from '../../../components/PageTitle';
+import LinkButton from '../../../components/LinkButton';
 
 interface WithRouterProps {
     router: NextRouter;
