+++ pages/admin/skills/add-skill.tsx
