+++ pages/admin/skills/index.tsx
@@ -2,11 +2,11 @@ import Link from 'next/link';
 import { Typography, Container, Grid } from '@mui/material';
 import { Box } from '@mui/system';
 import { useEffect, useState } from 'react';
-import LinkButton from '../../components/LinkButton';
-import EntityCard from '../../components/EntityCard';
-import AutoCompleteBox from '../../components/AutoCompleteBox';
-import PageTitle from '../../components/PageTitle';
-import ListGrid from '../../components/ListGrid';
+import LinkButton from '../../../components/LinkButton';
+import EntityCard from '../../../components/EntityCard';
+import AutoCompleteBox from '../../../components/AutoCompleteBox';
+import PageTitle from '../../../components/PageTitle';
+import ListGrid from '../../../components/ListGrid';
 
 export default function Skills() {
     interface MySkills {
