+++ pages/admin/training-methods/add-training-method.tsx
