+++ pages/admin/training-programmes/[id].tsx
@@ -3,8 +3,7 @@ import { useRouter } from 'next/router';
 import { useState, useEffect } from 'react';
 import { withRouter, NextRouter } from 'next/router'
 import DeleteForeverIcon from '@mui/icons-material/DeleteForever';
-import LinkButton from '../../components/LinkButton';
-import PageTitle from '../../components/PageTitle';
+import LinkButton from '../../../components/LinkButton';
 
 interface WithRouterProps {
     router: NextRouter
@@ -12,59 +11,32 @@ interface WithRouterProps {
 
 interface MyComponentProps extends WithRouterProps {}
 
-const EquineId: React.FC<MyComponentProps> = (props) => {
-    interface MyEquine {
-        category: {
-          id: number,
-          name: string
-        },
+const ProgrammeId: React.FC<MyComponentProps> = (props) => {
+    interface MyProgramme {
         id: number,
-        name: string,
-        programme: {
-          id: number,
-          name: string
-        },
-        skills: Array<any>,
-        yard: {
-          id: number,
-          name: string
-        }
+        name: string
       }
-    const [equine, setEquine] = useState<MyEquine>({
-        category: {
-          id: 0,
-          name: ""
-        },
+    const [programme, setProgramme] = useState<MyProgramme>({
         id: 0,
-        name: "",
-        programme: {
-          id: 0,
-          name: ""
-        },
-        skills: [],
-        yard: {
-          id: 0,
-          name: ""
-        }
+        name: ""
       });
 
     const router = useRouter();
 
-    const getEquineFromId = async () => {
-        const equineId = await props.router.query.id;
-        
-        await fetch(`${process.env.NEXT_PUBLIC_URL}/data/equines/${equineId}`)
+    const getProgrammeFromId = async () => {
+        const programmeId = await router.query.id;
+        await fetch(`${process.env.NEXT_PUBLIC_URL}/data/programmes/${programmeId}`)
             .then(response => response.json())
-            .then(data => setEquine(data))
+            .then(data => setProgramme(data))
             .catch(rejected => {
                 console.log(rejected);
             });
     }
 
-    const deleteEquineForever = async () => {
-        await fetch(`${process.env.NEXT_PUBLIC_URL}/data/equines/${equine.id}`, {method: 'DELETE'} )
+    const deleteProgrammeForever = async () => {
+        await fetch(`${process.env.NEXT_PUBLIC_URL}/data/programmes/${programme.id}`, {method: 'DELETE'} )
         .then(() => {
-            props.router.push('/equines')
+            props.router.push('/programmes')
         })
         .catch(rejected => {
             console.log(rejected);
@@ -72,7 +44,7 @@ const EquineId: React.FC<MyComponentProps> = (props) => {
     }
 
     useEffect(() => {
-        getEquineFromId();
+        getProgrammeFromId();
     }, []);
     
 
@@ -86,11 +58,6 @@ const EquineId: React.FC<MyComponentProps> = (props) => {
                     alignItems: 'center'
                 }}
             >
-                <PageTitle title={equine.name} sx={{ justifySelf: 'start' }}/>
-                <Card
-                    sx={{ my: '1rem', cursor: 'pointer', borderRadius: '20px' }}
-                >
-                </Card>
                 <Card
                     sx={{ my: '1rem', cursor: 'pointer', borderRadius: '20px' }}
                 >
@@ -100,28 +67,28 @@ const EquineId: React.FC<MyComponentProps> = (props) => {
                         gutterBottom
                         sx={{ my: '1rem', mx: '1rem' }}
                     >
-                        Yard: {equine.yard.name}
+                        Name: {programme.name}
                     </Typography>
                 </Card>
 
                 <Button
                     variant="outlined"
                     sx={{ my: '1rem' }}
-                    onClick={deleteEquineForever}
+                    onClick={deleteProgrammeForever}
                 >
                     <DeleteForeverIcon />
                 </Button>
 
                 <LinkButton
-                    buttonHref="/equines"
+                    buttonHref="/programmes"
                     variant="contained"
                     size="Large"
                     color="white"
-                    action="Go back to equines"
+                    action="Go back to Programmes"
                 />
             </Box>
         </Container>
     );
 };
 
-export default withRouter(EquineId);
+export default withRouter(ProgrammeId);
