+++ pages/admin/training-programmes/index.tsx
@@ -2,9 +2,9 @@ import Link from 'next/link';
 import { Typography, Container, Grid } from '@mui/material';
 import { Box } from '@mui/system';
 import { useEffect, useState } from 'react';
-import LinkButton from '../../components/LinkButton';
-import EntityCard from '../../components/EntityCard';
-import AutoCompleteBox from '../../components/AutoCompleteBox';
+import LinkButton from '../../../components/LinkButton';
+import EntityCard from '../../../components/EntityCard';
+import AutoCompleteBox from '../../../components/AutoCompleteBox';
 
 export default function Programmes() {
     interface MyProgrammes {
