+++ properties/properties.js
@@ -1,164 +1,164 @@
 const options = {
-    typeOfLearner : [
-        "Handled Youngster under 4 years",
-        "Unhandled Youngster under 4 years",
-        "Unhandled or semi-feral under 4 years",
-        "Unhandled or semi-feral over 4 years",
-        "Unhandleable, extremely fearful or known trauma",
-        "Handled on the farm",
-        "Previous skills but new to farm",
-        "Ridden potential",
-        "Ridden tickover",
-        "Companion Tickover / Maintenance",
-        "Sanctuary Tickover / Maintenance",
-        "Sactuary Frequent Handling"
-    ],
-    adminActions : [
-        {
-            title: 'Equines',
-            buttonTitle: 'Equine',
-            link: '/equines',
-            newLink: '/equines/add-equine',
-            editLink: '/equines/edit-equine',
-            deleteLink: '/equines/delete-equine',
-            newVariant: 'contained',
-            editVariant: 'outlined',
-            deleteVariant: 'outlined',
-            seeAllVariant: 'green',
-            newSize: 'large',
-            editSize: 'large',
-            deleteSize: 'small',
-            seeAllLink: '/equines'
-        },
-        {
-            title: 'Skills',
-            buttonTitle: 'Skill',
-            link: '/skills',
-            newLink: '/skills/add-skill',
-            editLink: '/skills/edit-skill',
-            deleteLink: '/skills/delete-skill',
-            newVariant: 'contained',
-            editVariant: 'outlined',
-            deleteVariant: 'outlined',
-            seeAllVariant: 'green',
-            newSize: 'large',
-            editSize: 'large',
-            deleteSize: 'small',
-            seeAllLink: '/skills'
-        },
-        {
-            title: 'Yards',
-            buttonTitle: 'Yard',
-            link: '/yards',
-            newLink: '/yards/add-yard',
-            editLink: '/yards/edit-yard',
-            deleteLink: '/yards/delete-yard',
-            newVariant: 'contained',
-            editVariant: 'outlined',
-            deleteVariant: 'outlined',
-            seeAllVariant: 'green',
-            newSize: 'large',
-            editSize: 'large',
-            deleteSize: 'small',
-            seeAllLink: '/yards'
-        },
-        {
-            title: 'Disruptions',
-            buttonTitle: 'Disruption',
-            link: '/disruptions',
-            newLink: '/disruptions/add-disruption',
-            editLink: '/disruptions/edit-disruption',
-            deleteLink: '/disruptions/delete-disruption',
-            newVariant: 'contained',
-            editVariant: 'outlined',
-            deleteVariant: 'outlined',
-            seeAllVariant: 'green',
-            newSize: 'large',
-            editSize: 'large',
-            deleteSize: 'small',
-            seeAllLink: '/disruptions'
-        },
-        {
-            title: 'Training Methods',
-            buttonTitle: 'Training Method',
-            link: '/training-methods',
-            newLink: '/training-methods/add-training-method',
-            editLink: '/training-methods/edit-training-method',
-            deleteLink: '/training-methods/delete-training-method',
-            newVariant: 'contained',
-            editVariant: 'outlined',
-            deleteVariant: 'outlined',
-            seeAllVariant: 'green',
-            newSize: 'large',
-            editSize: 'large',
-            deleteSize: 'small',
-            seeAllLink: '/training-methods'
-        },
-        {
-            title: 'Categories',
-            buttonTitle: 'Category',
-            link: '/categories',
-            newLink: '/categories/add-category',
-            editLink: '/categories/edit-category',
-            deleteLink: '/categories/delete-category',
-            newVariant: 'contained',
-            editVariant: 'outlined',
-            deleteVariant: 'outlined',
-            seeAllVariant: 'green',
-            newSize: 'large',
-            editSize: 'large',
-            deleteSize: 'small',
-            seeAllLink: '/categories'
-        },
-        {
-            title: 'Environments',
-            buttonTitle: 'Environment',
-            link: '/environments',
-            newLink: '/environments/add-environment',
-            editLink: '/environments/edit-environment',
-            deleteLink: '/environments/delete-environment',
-            newVariant: 'contained',
-            editVariant: 'outlined',
-            deleteVariant: 'outlined',
-            seeAllVariant: 'green',
-            newSize: 'large',
-            editSize: 'large',
-            deleteSize: 'small',
-            seeAllLink: '/environments'
-        },
-        {
-            title: 'Learner Types',
-            buttonTitle: 'Learner Type',
-            link: '/learner-types',
-            newLink: '/learner-types/add-learner-type',
-            editLink: '/learner-types/edit-learner-type',
-            deleteLink: '/learner-types/delete-learner-type',
-            newVariant: 'contained',
-            editVariant: 'outlined',
-            deleteVariant: 'outlined',
-            seeAllVariant: 'green',
-            newSize: 'large',
-            editSize: 'large',
-            deleteSize: 'small',
-            seeAllLink: '/learner-types'
-        },
-        {
-            title: 'Programmes',
-            buttonTitle: 'Programmes',
-            link: '/programmes',
-            newLink: '/programmes/add-programme',
-            editLink: '/programmes/edit-programme',
-            deleteLink: '/programmes/delete-programme',
-            newVariant: 'contained',
-            editVariant: 'outlined',
-            deleteVariant: 'outlined',
-            seeAllVariant: 'green',
-            newSize: 'large',
-            editSize: 'large',
-            deleteSize: 'small',
-            seeAllLink: '/programmes'
-        }
-    ]
-}
+	typeOfLearner: [
+		"Handled Youngster under 4 years",
+		"Unhandled Youngster under 4 years",
+		"Unhandled or semi-feral under 4 years",
+		"Unhandled or semi-feral over 4 years",
+		"Unhandleable, extremely fearful or known trauma",
+		"Handled on the farm",
+		"Previous skills but new to farm",
+		"Ridden potential",
+		"Ridden tickover",
+		"Companion Tickover / Maintenance",
+		"Sanctuary Tickover / Maintenance",
+		"Sactuary Frequent Handling",
+	],
+	adminActions: [
+		{
+			title: "Equines",
+			buttonTitle: "Equine",
+			link: "/admin/equines",
+			newLink: "/admin/equines/add-equine",
+			editLink: "/admin/equines/edit-equine",
+			deleteLink: "/admin/equines/delete-equine",
+			newVariant: "contained",
+			editVariant: "outlined",
+			deleteVariant: "outlined",
+			seeAllVariant: "green",
+			newSize: "large",
+			editSize: "large",
+			deleteSize: "small",
+			seeAllLink: "/equines",
+		},
+		{
+			title: "Skills",
+			buttonTitle: "Skill",
+			link: "/skills",
+			newLink: "/skills/add-skill",
+			editLink: "/skills/edit-skill",
+			deleteLink: "/admin/skills/delete-skill",
+			newVariant: "contained",
+			editVariant: "outlined",
+			deleteVariant: "outlined",
+			seeAllVariant: "green",
+			newSize: "large",
+			editSize: "large",
+			deleteSize: "small",
+			seeAllLink: "/skills",
+		},
+		{
+			title: "Yards",
+			buttonTitle: "Yard",
+			link: "/admin/yards",
+			newLink: "/admin/yards/add-yard",
+			editLink: "/admin/yards/edit-yard",
+			deleteLink: "/admin/yards/delete-yard",
+			newVariant: "contained",
+			editVariant: "outlined",
+			deleteVariant: "outlined",
+			seeAllVariant: "green",
+			newSize: "large",
+			editSize: "large",
+			deleteSize: "small",
+			seeAllLink: "/yards",
+		},
+		{
+			title: "Disruptions",
+			buttonTitle: "Disruption",
+			link: "/admin/disruptions",
+			newLink: "/admin/disruptions/add-disruption",
+			editLink: "/admin/disruptions/edit-disruption",
+			deleteLink: "/admin/disruptions/delete-disruption",
+			newVariant: "contained",
+			editVariant: "outlined",
+			deleteVariant: "outlined",
+			seeAllVariant: "green",
+			newSize: "large",
+			editSize: "large",
+			deleteSize: "small",
+			seeAllLink: "/disruptions",
+		},
+		{
+			title: "Training Methods",
+			buttonTitle: "Training Method",
+			link: "/training-methods",
+			newLink: "/admin/training-methods/add-training-method",
+			editLink: "/admin/training-methods/edit-training-method",
+			deleteLink: "/admin/training-methods/delete-training-method",
+			newVariant: "contained",
+			editVariant: "outlined",
+			deleteVariant: "outlined",
+			seeAllVariant: "green",
+			newSize: "large",
+			editSize: "large",
+			deleteSize: "small",
+			seeAllLink: "/training-methods",
+		},
+		{
+			title: "Categories",
+			buttonTitle: "Category",
+			link: "/admin/categories",
+			newLink: "/admin/ategories/add-category",
+			editLink: "/admin/categories/edit-category",
+			deleteLink: "/admin/categories/delete-category",
+			newVariant: "contained",
+			editVariant: "outlined",
+			deleteVariant: "outlined",
+			seeAllVariant: "green",
+			newSize: "large",
+			editSize: "large",
+			deleteSize: "small",
+			seeAllLink: "/categories",
+		},
+		{
+			title: "Environments",
+			buttonTitle: "Environment",
+			link: "/admin/environments",
+			newLink: "/admin/environments/add-environment",
+			editLink: "/admin/environments/edit-environment",
+			deleteLink: "/admin/environments/delete-environment",
+			newVariant: "contained",
+			editVariant: "outlined",
+			deleteVariant: "outlined",
+			seeAllVariant: "green",
+			newSize: "large",
+			editSize: "large",
+			deleteSize: "small",
+			seeAllLink: "/environments",
+		},
+		{
+			title: "Learner Types",
+			buttonTitle: "Learner Type",
+			link: "/admin/learner-types",
+			newLink: "/admin/learner-types/add-learner-type",
+			editLink: "/admin/learner-types/edit-learner-type",
+			deleteLink: "/admin/learner-types/delete-learner-type",
+			newVariant: "contained",
+			editVariant: "outlined",
+			deleteVariant: "outlined",
+			seeAllVariant: "green",
+			newSize: "large",
+			editSize: "large",
+			deleteSize: "small",
+			seeAllLink: "/learner-types",
+		},
+		{
+			title: "Programmes",
+			buttonTitle: "Programmes",
+			link: "/admin/programmes",
+			newLink: "/admin/programmes/add-programme",
+			editLink: "/admin/programmes/edit-programme",
+			deleteLink: "/admin/programmes/delete-programme",
+			newVariant: "contained",
+			editVariant: "outlined",
+			deleteVariant: "outlined",
+			seeAllVariant: "green",
+			newSize: "large",
+			editSize: "large",
+			deleteSize: "small",
+			seeAllLink: "/programmes",
+		},
+	],
+};
 
 export default options;
